Add CSS styles and minor content changes
The bulk of this update was adding styles.css and the styling changes.
Return to homepage link was moved back to the top, for aesthetics.
Minor content changes for formatting purposes prior to flexbox lessons.

Introduced first BEM class styling for hero-section.
Prepared stylesheet for tokens and replacement.

to-do: add more wrappers for content alignment around lists and
descriptions. replace with tokens where applicable. Implement
media queries and text scaling if TOP covers it before moving on.